requirements.txt
fastapi
langchain-core
langchain-huggingface
huggingface-hub
sentence_transformers 
langchain_community
langchain 
faiss-cpu
torch
pypdf
python-multipart
python-dotenv
cohere
uvicorn